server: handle the case that a user tries to invite an user who already exists
This will be a rather common case if a lot of users exists.

Instead of letting the database throw an error the database should ignore
the insert request. We then can the detect this as now row has been
inserted (rowCount !== 1).

For now just do nothing.

We should return a message to the client. The client then can then notify the
user via the page status.