use new acceped column
1. if a user declined a membership and accepted is therefor false, do not
   list this group any more for him
2. if a user as not (yet) accepted a membership and accpeted is therefor
   false (user declined membership) or the request is still pending (NULL)
   let members of the group not see any information of the user beyond his
   uid.
   Of course: if both user habe another common group this information is
   know by those clients and will be displayed.

If a user accepts an invitation to a group, the group_membership table will
get an update and all clients connected clients will get updates for the
memberships. They then will detect they miss the user information (if they
do not have it already because of other groups they share).

To make ordering of updattes easier, always send last_mod_seq for reach row.
The client cache class does not use it yet but will so to avoid to update
entries with older information.